Address PR review findings for GC9A01 display usermod
- Honor showHowLong in overlay(): overlayUntil now uses the caller-supplied
  duration when > 0, falling back to overlayInactivityTimeout — fixes short
  rotary overlays and displayNetworkInfo() being held too long
- Guard wakeDisplayFromSleep() with displayEnabled check so backlight/state
  are not restored when the usermod is disabled
- Change displayTimeout from uint16_t to uint32_t to hold values up to
  300 000 ms without overflow
- Wrap GC9A01-specific wakeDisplay()/redraw() calls in rotary encoder ALT
  with #ifdef USERMOD_GC9A01_DISPLAY to prevent void* dereference when
  the display type is not compiled in
## Bug Fixes (PR review follow-up)
- `overlay()`: honor `showHowLong` parameter — overlayUntil now respects explicit durations (e.g. from `displayNetworkInfo()`) instead of always using `overlayInactivityTimeout`
- `wakeDisplayFromSleep()`: skip backlight restore and state reset when `displayEnabled` is false
- `displayTimeout`: widened from `uint16_t` to `uint32_t` to safely hold values up to 300 000 ms
- Rotary encoder ALT: GC9A01 `wakeDisplay()`/`redraw()` calls now guarded by `#ifdef USERMOD_GC9A01_DISPLAY`

+++ usermods/usermod_v2_gc9a01_display/usermod_v2_gc9a01_display.h
@@ -50,7 +50,7 @@ class UsermodGC9A01Display : public Usermod {
     bool showingClock = false; // Track if currently showing clock (for rotary encoder state reset)
     unsigned long welcomeScreenStartTime = 0;
     uint8_t backlight = 75; // Backlight brightness percentage (0-100%, default 75%)
-    uint16_t displayTimeout = 60000; // 60 seconds default
+    uint32_t displayTimeout = 60000; // 60 seconds default
     bool sleepMode = true; // Enable sleep mode by default
     bool clockMode = false; // Show clock only when idle
     bool flip = false; // Display rotation (0 or 2)
